Improve maintainability boundaries
