+++ docs/adr/0051-adopt-shared-contracts-and-logging-foundations.md
@@ -1,7 +1,7 @@
 ---
 id: 0051
 title: Adopt shared contracts and logging foundations
-status: Proposed
+status: Accepted
 date: 2026-05-11
 supersedes: []
 superseded-by: null
@@ -16,6 +16,7 @@ affects:
   - src/contracts/explorer.ts
   - src/contracts/knowledgeBase.ts
   - src/contracts/memory.ts
+  - src/contracts/responses.ts
   - src/contracts/serviceTier.ts
   - src/contracts/settings.ts
   - src/contracts/streams.ts
@@ -37,27 +38,42 @@ affects:
   - src/routes/chat/workspaceInstructionRoutes.ts
   - src/services/chatService.ts
   - src/services/chat/attachments.ts
+  - src/services/chat/artifactStore.ts
   - src/services/chat/messageQueueStore.ts
+  - src/services/chat/usageLedgerStore.ts
+  - src/services/chat/workspaceFeatureSettingsStore.ts
   - src/services/chat/workspaceInstructionStore.ts
+  - src/services/contextMap/autoApply.ts
+  - src/services/contextMap/candidatePrimitives.ts
   - src/services/contextMap/service.ts
   - src/services/contextMap/jsonRepair.ts
   - src/services/contextMap/pipelineMetadata.ts
+  - src/services/contextMap/sourcePlanning.ts
   - src/services/memoryWatcher.ts
   - src/services/streamJobSupervisor.ts
   - src/utils/logger.ts
   - src/ws.ts
   - mobile/AgentCockpitPWA/src/api.ts
+  - mobile/AgentCockpitPWA/src/appModel.ts
+  - mobile/AgentCockpitPWA/src/useViewportHeightVar.ts
   - web/AgentCockpitWeb/src/api.js
   - web/AgentCockpitWeb/src/shell.jsx
+  - web/AgentCockpitWeb/src/shellState.jsx
   - web/AgentCockpitWeb/src/streamStore.js
   - web/AgentCockpitWeb/src/chat/attachments.jsx
   - web/AgentCockpitWeb/src/chat/messageParsing.ts
   - web/AgentCockpitWeb/src/chat/queue.jsx
   - test/chatContracts.test.ts
+  - test/chat.artifactStore.test.ts
+  - test/chat.usageLedgerStore.test.ts
+  - test/chat.workspaceFeatureSettingsStore.test.ts
   - test/contextMap.jsonRepair.test.ts
   - test/contextMap.pipelineMetadata.test.ts
   - test/frontendMessageParsing.test.ts
+  - test/mobileAppModel.test.ts
   - test/logger.test.ts
+  - scripts/check-maintainability.js
+  - scripts/check-spec-drift.js
   - docs/spec-api-endpoints.md
   - docs/spec-backend-services.md
   - docs/spec-context-map.md
@@ -94,22 +110,28 @@ Adopt small shared foundations for maintainability work:
 - Keep `src/routes/chat.ts` as the composition root for chat dependencies and
   stream orchestration, but move focused route groups into
   `src/routes/chat/*Routes.ts`.
-- Keep `ChatService` as the public facade while moving pure helpers into
-  focused modules under `src/services/chat/`, including private persistence
-  stores such as message queue storage and workspace-instruction compatibility
-  storage.
-- Keep Context Map processor orchestration in `service.ts`, but move local JSON
-  extraction/repair helpers into `src/services/contextMap/jsonRepair.ts` and
-  pure run/synthesis metadata helpers into
-  `src/services/contextMap/pipelineMetadata.ts`.
-- Put pure V2 chat-message parsing helpers and large V2 chat UI subtrees under
-  `web/AgentCockpitWeb/src/chat/` instead of inside `shell.jsx`.
+- Keep `ChatService` as the public facade while moving pure helpers and private
+  persistence into focused modules under `src/services/chat/`, including
+  artifact storage, usage ledger storage, workspace feature settings, message
+  queue storage, and workspace-instruction compatibility storage.
+- Keep Context Map processor orchestration in `service.ts`, but move focused
+  work into owned modules under `src/services/contextMap/`: JSON
+  extraction/repair, run/synthesis metadata, source planning, candidate
+  normalization primitives, and safe auto-apply policy.
+- Put pure V2 chat-message parsing helpers, large V2 chat UI subtrees, and
+  shell state providers under owned web modules instead of keeping all desktop
+  UI behavior inside `shell.jsx`.
+- Put mobile projection/parsing helpers and viewport sizing behavior under
+  owned PWA modules instead of keeping all mobile behavior inside `App.tsx`.
+- Add stable maintainability checks for browser-safe contract boundaries, web
+  and mobile server-import boundaries, backend structured logging migration,
+  and chat route/spec drift.
 - Use `src/utils/logger.ts` as the structured logging entrypoint for migrated
   backend code, with `LOG_LEVEL` filtering, metadata redaction, bounded rich
   value serialization, and cycle-safe metadata handling.
 
-These foundations are allowed to start narrow. New code should prefer them when
-touching the same boundary, but existing modules can migrate in focused slices.
+These boundaries define ownership for the completed slices and for similar
+changes that touch the same behavior.
 
 ## Alternatives Considered
