+++ docs/spec-frontend.md
@@ -21,6 +21,7 @@ The sidebar footer reads `GET /api/me` and supports first-party local owner sess
 - `web/AgentCockpitWeb/index.html` — Vite HTML entry. It keeps the root element, pre-paint theme script, font links, and highlight.js theme stylesheet, then imports `/src/main.jsx` as a module. It does not load React, ReactDOM, Babel Standalone, marked, DOMPurify, or highlight.js from CDNs; those are bundled from root npm dependencies.
 - `web/AgentCockpitWeb/src/main.jsx` — Vite module entry. It imports the CSS files, imports `initTabIndicator`, imports `shell.jsx`, then initializes the tab indicator. App-local frontend dependencies are resolved through ES module imports/exports instead of the retired Browser-Babel `window.*` compatibility bridge. The generated HTML contains hashed `/v2/assets/*` references, so manual `?v=N` cache-busters are no longer part of the V2 runtime.
 - `web/AgentCockpitWeb/src/syntaxHighlight.js` — highlight.js registry module. It imports `highlight.js/lib/core`, registers the language set the chat renderer uses (`bash`, shell aliases, CSS, diff, JavaScript/JSX, JSON, Markdown, Python, SQL, TypeScript/TSX, HTML/XML, YAML), and exports the configured `hljs` instance for `shell.jsx`.
+- `web/AgentCockpitWeb/src/shellState.jsx` — shell-level React contexts/hooks for backend metadata, CLI profiles, CLI update snapshots, per-conversation stream-store subscriptions, sidebar stream-state subscriptions, backend inline icons, assistant avatars, and lazy-screen loading chrome. `shell.jsx` imports these focused shell utilities and keeps route/view orchestration local.
 - `web/AgentCockpitWeb/src/fileLinks.ts` — typed helper for resolving assistant Markdown links into safe local file references. It accepts absolute workspace paths and active-conversation `data/chat/artifacts/{convId}/{filename}` paths, strips optional `:line` / `:line:column` suffixes, decodes URL-encoded `file://` paths, and rejects web URLs, relative paths, traversal-looking paths, nested artifact paths, and artifacts from other conversations.
 - `web/AgentCockpitWeb/src/chat/attachments.jsx` — composer attachment tray and typed attachment chip subtree. It owns file kind badges, upload progress display, OCR/dissolve affordances, image blob-preview lifecycle, and restored-upload thumbnails. The shell imports `AttTray` from here instead of carrying attachment rendering inline.
 - `web/AgentCockpitWeb/src/chat/messageParsing.ts` — pure helper module for chat-message marker parsing. It extracts hidden `<!-- FILE_DELIVERY:/abs/path -->` markers, parses `[Uploaded files: ...]` tags into paths, and centralizes stream-error message ids that should stay hidden from the rendered feed.
