+++ docs/spec-mobile-pwa.md
@@ -33,6 +33,8 @@ See [ADR-0025](adr/0025-use-mobile-pwa-as-sole-mobile-client.md).
 - `public/manifest.webmanifest`, `public/icon.svg`, `public/icon-192.png`, `public/icon-512.png`, and `public/apple-touch-icon.png` — tracked source install metadata and icons copied into `public/mobile-built/` during build. iOS home-screen installs use the 180x180 PNG `apple-touch-icon`; the manifest also lists PNG icons because Safari is not reliable with SVG-only PWA icons.
 - `src/main.tsx` — React entrypoint.
 - `src/App.tsx` — mobile UI shell and state machine. It renders conversation list, chat transcript, composer, queue stack/editor, run settings, conversation actions, sessions, workspace files, file preview/share, attachment tray, interaction cards, and stream controls.
+- `src/appModel.ts` — browser-only mobile app model helpers for file-delivery marker parsing, typed file preview references, reset-session list reconciliation, conversation-list projection, queue wire content, attachment filtering, workspace path formatting, and shared formatters. These helpers are covered separately so the main UI shell can stay focused on state orchestration.
+- `src/useViewportHeightVar.ts` — visual-viewport hook that owns the iOS Safari app-shell sizing variables (`--app-height`, `--app-width`, `--app-top`, and `--app-left`) and document scroll reset behavior.
 - `src/api.ts` — same-origin TypeScript API client for current-user, settings, backends, profile metadata, conversations, streams, queue, sessions, attachments/OCR, file delivery, workspace explorer, and WebSocket URLs. Conversation create, message send, stdin input, and explorer mutation bodies use type-only imports from browser-safe `src/contracts/*` files so mobile request shapes compile against the same contract types as the server. State-changing requests lazily fetch CSRF tokens and send `x-csrf-token`. Multipart uploads use `XMLHttpRequest` for progress and cancel.
 - `src/types.ts` — TypeScript mirrors for the PWA data models.
 - `src/styles.css` — mobile-first CSS for the PWA shell.
