+++ package.json
@@ -17,6 +17,8 @@
     "mobile:dev": "npm --prefix mobile/AgentCockpitPWA run dev",
     "mobile:build": "npm --prefix mobile/AgentCockpitPWA run build",
     "mobile:typecheck": "npm --prefix mobile/AgentCockpitPWA run typecheck",
+    "maintainability:check": "node scripts/check-maintainability.js",
+    "spec:drift": "node scripts/check-spec-drift.js",
     "adr:new": "node scripts/adr-new.js",
     "adr:index": "node scripts/adr-index.js",
     "adr:lint": "node scripts/adr-lint.js"
