+++ src/contracts/chat.ts
@@ -1,5 +1,12 @@
-import type { QueuedMessage, Settings } from '../types';
-import { asRecord, contractError } from './validation';
+import type { Settings } from '../types';
+import type {
+  BackendMetadata,
+  CliUpdatesResponse,
+  QueuedMessage,
+  UpdateStatus,
+  UsageLedger,
+} from './responses';
+import { asRecord, contractError, optionalString, optionalStringEnum } from './validation';
 export { parseServiceTierInput } from './serviceTier';
 
 export interface ApiErrorResponse {
@@ -11,6 +18,26 @@ export interface VersionResponse {
   remoteVersion: string | null;
   updateAvailable: boolean;
 }
+
+export interface BasicOkResponse {
+  ok: boolean;
+}
+
+export interface CurrentUserResponse {
+  displayName: string | null;
+  email: string | null;
+  provider: 'local' | 'google' | 'github' | null;
+}
+
+export interface BackendsResponse {
+  backends: BackendMetadata[];
+}
+
+export interface UpdateStatusResponse extends UpdateStatus {}
+
+export interface CliUpdatesEnvelope extends CliUpdatesResponse {}
+
+export interface UsageStatsResponse extends UsageLedger {}
 
 export interface QueueResponse {
   queue: QueuedMessage[];
@@ -37,10 +64,39 @@ export function validateQueueUpdateRequest(body: unknown): QueueUpdateRequest {
 }
 
 export function validateSettingsRequest(body: unknown): Settings {
-  if (!body || typeof body !== 'object' || Array.isArray(body)) {
-    contractError('settings must be an object');
+  const record = asRecord(body, 'settings must be an object');
+  optionalStringEnum(record, 'theme', ['light', 'dark', 'system'], 'theme must be light, dark, or system');
+  optionalStringEnum(record, 'sendBehavior', ['enter', 'ctrlEnter'], 'sendBehavior must be enter or ctrlEnter');
+  optionalString(record, 'systemPrompt', 'systemPrompt must be a string');
+  optionalString(record, 'defaultBackend', 'defaultBackend must be a string');
+  optionalString(record, 'defaultCliProfileId', 'defaultCliProfileId must be a string');
+  optionalString(record, 'defaultModel', 'defaultModel must be a string');
+  optionalStringEnum(record, 'defaultEffort', ['none', 'minimal', 'low', 'medium', 'high', 'xhigh', 'max'], 'defaultEffort is invalid');
+  optionalStringEnum(record, 'defaultServiceTier', ['fast'], 'defaultServiceTier is invalid');
+  optionalString(record, 'workingDirectory', 'workingDirectory must be a string');
+  validateCliProfiles(record.cliProfiles);
+  validateOptionalObject(record.memory, 'memory');
+  validateOptionalObject(record.knowledgeBase, 'knowledgeBase');
+  validateOptionalObject(record.contextMap, 'contextMap');
+  return record as unknown as Settings;
+}
+
+function validateCliProfiles(value: unknown): void {
+  if (value === undefined) return;
+  if (!Array.isArray(value)) contractError('cliProfiles must be an array');
+  for (const profile of value) {
+    const record = asRecord(profile, 'cliProfiles entries must be objects');
+    optionalString(record, 'id', 'cliProfiles entries must have string ids');
+    optionalString(record, 'name', 'cliProfiles entries names must be strings');
+    optionalString(record, 'vendor', 'cliProfiles entries vendors must be strings');
   }
-  return body as Settings;
+}
+
+function validateOptionalObject(value: unknown, name: string): void {
+  if (value === undefined) return;
+  if (!value || typeof value !== 'object' || Array.isArray(value)) {
+    contractError(`${name} must be an object`);
+  }
 }
 
 function validateQueuedMessage(entry: unknown): void {
