+++ src/contracts/streams.ts
@@ -1,6 +1,7 @@
 import { parseServiceTierInput, type ContractServiceTier, type ServiceTierInput } from './serviceTier';
-import { asRecord, optionalBoolean, optionalString, requiredNonEmptyString } from './validation';
-import type { ContractEffortLevel } from './conversations';
+import { asRecord, optionalBoolean, optionalString, optionalStringEnum, requiredNonEmptyString } from './validation';
+import type { Message, StreamJobRuntimeInfo } from './responses';
+import { CONTRACT_EFFORT_LEVELS, type ContractEffortLevel } from './conversations';
 
 export interface SendMessageRequest {
   content: string;
@@ -20,13 +21,40 @@ export interface ConversationInputRequest {
   streamActive?: boolean;
 }
 
+export interface SendMessageResponse {
+  userMessage: Message;
+  streamReady: boolean;
+}
+
+export interface ConversationInputResponse {
+  mode: 'stdin' | 'message';
+}
+
+export interface ActiveStreamResponse {
+  id: string;
+  jobId?: string | null;
+  state?: string;
+  backend: string;
+  startedAt: string | null;
+  lastEventAt: string | null;
+  connected: boolean;
+  runtimeAttached: boolean;
+  pending: boolean;
+  runtime: StreamJobRuntimeInfo | null;
+}
+
+export interface ActiveStreamsResponse {
+  ids: string[];
+  streams: ActiveStreamResponse[];
+}
+
 export function validateSendMessageRequest(body: unknown): ValidatedSendMessageRequest {
   const record = asRecord(body);
   return {
     content: requiredNonEmptyString(record, 'content', 'Message content required'),
     backend: optionalString(record, 'backend'),
     model: optionalString(record, 'model'),
-    effort: optionalString(record, 'effort') as ContractEffortLevel | undefined,
+    effort: optionalStringEnum(record, 'effort', CONTRACT_EFFORT_LEVELS),
     cliProfileId: optionalString(record, 'cliProfileId'),
     serviceTier: parseServiceTierInput(record.serviceTier),
   };
